more async refactoring

+++ GlasgowAstro.GuideAlert/Models/GuideAlertSettings.cs
@@ -12,6 +12,6 @@
 
         public string PhdHost { get; set; }
 
-        public ushort PhdPort { get; set; }
+        public int PhdPort { get; set; }
     }
 }
